backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.5.2
firebase-admin==6.4.0
Pillow==10.2.0
pydantic==2.6.3
python-multipart==0.0.9
